% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.acvfilter.R
\name{check.acvfilter}
\alias{check.acvfilter}
\title{Chequea el acv y el filter}
\usage{
check.acvfilter(awdfile, set)
}
\arguments{
\item{awdfile}{que seria un string con el nombre del archivo sin la extension}

\item{set}{el objeto de settings}
}
\value{
una lista con varias cosas, algun dia actualizare esta parte
}
\description{
Esta funcion mantiene actualizada el archivo acv (actividad) y filtro para cada sujeto
}
\examples{
# awdfile <- "2086-308-045 CHT Visit2.AWD"
# acveditRDS <- check.acvfilter(awdfile, set)
}
