% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_function_check.epidata.R
\name{check.epidata}
\alias{check.epidata}
\title{Checador de archivo EPI}
\usage{
check.epidata(data = NULL)
}
\arguments{
\item{data}{Es el objeto data.frame proveniente de la funcion validEvents}
}
\value{
Nada, solo informacion
}
\description{
Esta funcion entrega guias de como debe estar el archivo EPI
para pase por las funciones de analisis. Ojo que ya debiera venir medio
modificado desde la funcion validEvents
}
\examples{
# Revisar como debiera ser el archivo epi
# check.epidata()
#
# Revisar un epi preprocesado
# check.epidata(epi)
}
