% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.actogram.R
\name{create.actogram}
\alias{create.actogram}
\title{Crear actograma}
\usage{
create.actogram(semiper, set, filterRDS, fy = 1)
}
\arguments{
\item{semiper}{corresponde al resutlado de la funcion check.acvfilter en especial el elemento 'semiper'}

\item{set}{Settings file}

\item{filterRDS}{es la lista del poll de fitro que carga la app}

\item{fy}{es el factor de expansion de y}
}
\value{
un grafico
}
\description{
Toma plot simple y crea un actogama
}
\examples{
# setwd("D:/OneDrive/INTA/Actigrafia/testfolder")
# set <- getset(getwd())
# filterRDS <- readRDS("2058-001-368 JRG Baseline.edit.RDS")
# acveditRDS <- check.acvfilter("2058-001-368 JRG Baseline.AWD", set)
# acveditRDS <- acveditRDS$semiper
# windows()
# create.actogram(acveditRDS, set, filterRDS)
}
