% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.acv.R
\name{create.acv}
\alias{create.acv}
\title{Crea el primer ACV}
\usage{
create.acv(awdfile, set)
}
\arguments{
\item{awdfile}{es el string del archivo awd (el original)}

\item{set}{El objeto de settings}
}
\value{
Un data frame con toda la data necesaria para comenzar el proceso
}
\description{
ACV es Activity, esta funcion toma el raw data y hace la primera pasada de computos. Crea la secuencia, quita los marcadores M, aplica el algoritmo del minimitter y el estado actigrafico suavizado segun el valor de consolidacion de estado. Hace varios recortes que ya documentare.
}
\examples{
# act <- create.acv("2086-308-045 CHT Visit2.AWD", sensi = set$sensivar)
}
