% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.acvedit.R
\name{create.acvedit}
\alias{create.acvedit}
\title{Crea archivo edicion de actividad}
\usage{
create.acvedit(awdfile, acv.filter, filter.stats)
}
\arguments{
\item{awdfile}{el string del nobmre de archivo}

\item{acv.filter}{es un data frame ACV}

\item{filter.stats}{es el resultado de la funcion create.firstfilter}
}
\value{
el acv con los filtros actualizados y ademas guarda el archivo en RDS
}
\description{
Esta funcion toma un awdfile lo que sale de create.acv y el filtro para hacer un data frame con la data etiquetada en al variable filtro, basicamente es un dataframe acv que tiene actualizado los filtros.
}
\examples{
# setwd("D:/OneDrive/INTA/Actigrafia/testfolder")
# acv <- create.acv("2086-308-045 CHT Visit2.AWD", sensi = set$sensivar)
# semiper <- create.semiper(acv)
# filtros <- create.firstfilter("2086-308-045 CHT Visit2.AWD", semiper)
# acvEdit <- create.acvedit("2086-308-045 CHT Visit2.AWD", acv, filtros)
}
