% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.epi.R
\name{create.epi}
\alias{create.epi}
\title{Crea archivo de episodios EPI}
\usage{
create.epi(acvedit = NULL, filter = NULL, set = NULL, dia0 = FALSE)
}
\arguments{
\item{acvedit}{el objeto que sale de check.acvfilter}

\item{filter}{es el archivo de filtro guardado RDS}

\item{set}{es la lista de settings del archivo sleep.config}

\item{dia0}{logico que indica si se conserva el dia 0 o no en el epi final}
}
\value{
data frame
}
\description{
Este es el archivo master que usa todo para crear los episodios con los cuales se calculan las estadisticas.
}
\examples{
# setwd("D:/OneDrive/INTA/Actigrafia/testfolder")
# set <- getset(getwd())
# awdfile <- "2058-001-368 JRG Baseline"
# acvedit <- check.acvfilter(awdfile, set)
# filter <- readRDS(paste0(awdfile, ".edit.RDS"))
}
