% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.firstfilter.R
\name{create.firstfilter}
\alias{create.firstfilter}
\title{crea el primer filtro}
\usage{
create.firstfilter(awdfile, semiper)
}
\arguments{
\item{awdfile}{el string de nombre de archivo}

\item{semiper}{el archivo de semiperiodos que sale de create.semiper(acv)}
}
\value{
Devuelte un data frame y un rds al disco
}
\description{
toma un acv y en virtud de la actividad determina los primeros filtros que existen de forma automatica y ademas guarda el rds
}
\examples{
# acv <- create.acv("2086-308-045 CHT Visit2.AWD", set)
# semiper <- create.semiper(acv, set)
# filtros <- create.firstfilter("2086-308-045 CHT Visit2.AWD", semiper)
}
