% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plotActo.R
\name{create.plotActo}
\alias{create.plotActo}
\title{Crea plot de edicion}
\usage{
create.plotActo(gdata, set, filterRDS, pct.y = 1)
}
\arguments{
\item{gdata}{el periodo que sale de check.acvedit}

\item{set}{el objeto de settings}

\item{filterRDS}{lista del poll de los filtros en la app}

\item{pct.y}{el valor de ponderacion del eje y}
}
\value{
un plot
}
\description{
Crea un plot para cada periodo todo lindo para la parte de la edicion
}
\examples{
# setwd("D:/OneDrive/INTA/Actigrafia/testfolder")
# set <- getset(getwd())
# acveditRDS <- check.acvfilter("2058-002-298 MLR Baseline.AWD", set)
# acveditRDS <- acveditRDS$semiper
# filterRDS <- readRDS("2058-002-298 MLR Baseline.edit.RDS")
# gdata <- acveditRDS[["per00"]]
# create.plotActo(gdata, set, filterRDS)
}
