% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plotSimple.R
\name{create.plotSimple}
\alias{create.plotSimple}
\title{Plot simple para periodo.}
\usage{
create.plotSimple(gdata, set, filterRDS, pct.y = 1, limites = NULL, lw = 1)
}
\arguments{
\item{gdata}{datos que saca el check.acvfilter}

\item{set}{Settings file}

\item{filterRDS}{lista del poll de los filtros en la app}

\item{pct.y}{Factor de multiplicacion para el eje y}

\item{limites}{Vector de limites para el eje x}

\item{lw}{Numeric de weight para las bandas del grafico}
}
\value{
plot de un periodo
}
\description{
Hace un plot individual para un periodo en version simple para el actograma
}
\examples{
# setwd("D:/OneDrive/INTA/Actigrafia/testfolder")
# set <- getset(getwd())
# acveditRDS <- check.acvfilter("2058-001-368 JRG Baseline.AWD", set)
# acveditRDS <- acveditRDS$semiper
# filterRDS <- readRDS("2058-001-368 JRG Baseline.edit.RDS")
# gdata <- acveditRDS[["per08"]]
# create.plotSimple(gdata, set, filterRDS)
}
