% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.semiper.R
\name{create.semiper}
\alias{create.semiper}
\title{Crear semi periodos}
\usage{
create.semiper(acv, set)
}
\arguments{
\item{acv}{es el data frame de la funcion create.acv}

\item{set}{es el setting object}
}
\value{
una lista con todos los semiperiodos
}
\description{
Un periodo es noche + dia, lo que hace es separar el acv y lo deja listo para hacer graficos y eso.
}
\examples{
# setwd("D:/OneDrive/INTA/Actigrafia/testfolder")
# acv <- create.acv("2086-308-045 CHT Visit2.AWD", sensi = set$sensivar)
# semiper <- create.semiper(acv)
# names(semiper)
# [1] "d0" "d1" "d2" "d3" "d4" "d5" "d6" "d7" "n1" "n2" "n3" "n4" "n5" "n6" "n7"
}
