% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.segment.R
\name{find.segment}
\alias{find.segment}
\title{find.segment}
\usage{
find.segment(df = NULL, var = NULL, filtro = NULL)
}
\arguments{
\item{df, }{un data.frame}

\item{var, }{varible donde buscar}

\item{filtro, }{un valor que buscar}
}
\value{
data.frame con el inicio y fin (en index) de la posicion del filtro
}
\description{
Encuentra segmentos de algo en un vetor
}
\examples{
# find.segment(mtcars, "am", 1)

}
