% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_function_24h.R
\name{function_24h}
\alias{function_24h}
\title{Analisis de tipo 24h}
\usage{
function_24h(epi)
}
\arguments{
\item{epi}{data.frame de epi tipo validEvents}
}
\value{
data.frame con los resultados
}
\description{
Funcion para analizar data de 24 horas
24 horas se entiende como la presencia de períodos consecutivos.
calcula latencias como se describe en la documentacion. v1 20.06.2020.
}
\examples{
# test <- function_24h(epi) 
#

}
