% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_function_combi24h.R
\name{function_combi24h}
\alias{function_combi24h}
\title{Combinar peridoso para formar 24h}
\usage{
function_combi24h(epi)
}
\arguments{
\item{epi}{data.frame de epi tipo validEvents}
}
\value{
data.frame con los resultados
}
\description{
Funcion para COMBINACIONES de 24 horas, para analisis de causa efecto
v1 22.06.2020 - 24 horas se entiende como la presencia de períodos consecutivos.
calcula latencias como se describe en la documentacion.
}
\examples{
# test <- function_combi24h(epi) 
#

}
