% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_function_conteo.R
\name{function_conteo}
\alias{function_conteo}
\title{Calculo de cantidades o conteos de episodios para cada periodo}
\usage{
function_conteo(epi)
}
\arguments{
\item{epi}{es el data.frame que arroja el validEvents}
}
\value{
data.frame con los analisis
}
\description{
Funcion para capturar N de Siestas y Despertares para cada semi
periodo en epi.data v1.0 14.03.2019. Usa un epi data. Hace el calculo para
un semi periodo y asigna siesta en dia y despertar en noche, tambien si es
posible hace asignaci?n a mitad o tercio, no usa la data expandida, en su
lugar  expande luego de filtrar el suejto.
}
\examples{
# conteo <- function_conteo(epi)
#
}
