% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_function_maxdur.R
\name{function_duracionMax}
\alias{function_duracionMax}
\title{Capturar el episodio mas largo de un periodo}
\usage{
function_duracionMax(epi)
}
\arguments{
\item{epi}{data.frame de epi tipo validEvents}
}
\value{
data.frame con los resultados
}
\description{
Captura el episodio mas largo en cada periodo (Dia o Noche)
para sueno y vigilia - v1.0 11.03.2019.
Usa un epi data. hay que usar data expandida para calcular a que parte
de la noche pertenece, se puede mas eficiente pero hay que hacerlo   :)
}
\examples{
# function_duracionMax(epi)
#
}
