% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_function_duration.R
\name{function_duration}
\alias{function_duration}
\title{Estadisticas de duracion de episodios en periodos}
\usage{
function_duration(epi)
}
\arguments{
\item{epi}{data.frame preprocesado con el validEvents}
}
\value{
data.frame con los analisis
}
\description{
Funcion para capturar Duracion de Siestas y Despertares para
cada semi periodo en epi.data v1.0 01.04.2019. Usa un epi data. Hace el
calculo para un semi periodo (solo dia o noche) y calcula el tiempo de vigilia o sueno  segun
corresponda y asigna a todo, mitad y tercio.  Para ello usa la data
expandida y lo hace de por sujeto. Futuro desarrollador... si se que la
función es lenta, pero si uso la librería data.table caga todo porque se
superpone a dplyr y funciones del base y adaptar me dio una laaaaaaaata enorme.
}
\examples{
# function_duracion(epi)
#
#
}
