% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_function_expand.epidata.R
\name{function_expand.epidata}
\alias{function_expand.epidata}
\title{Funcion de expansion de datos}
\usage{
function_expand.epidata(data)
}
\arguments{
\item{data}{es un epi del preproceso validEvents}
}
\value{
data.frame con los resultados
}
\description{
para hacer la expansion de episodios de sue?o o vigilia
basandose en duracion y la hora de inicio del mismo 11.03.2019. Transforma
un archvio tipo epi (por episodio) a uno por minuto, si se que podria usar
directamente un acvdata pero esta solucion es pervia a la creacion de la
aplicacion. So suck it up...   :P   .
}
\examples{
# function_expand.epidata(epi)
#
}
