% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_function_latencia.R
\name{function_latencia}
\alias{function_latencia}
\title{Latencia al primer W}
\usage{
function_latencia(epi)
}
\arguments{
\item{epi}{data.frame de epi tipo validEvents}
}
\value{
data.frame con los resultados
}
\description{
Funcion para capturar la latencia al primer W y al 2do si es que
hay, ademas del tiempo 2do sueno si es que hay - v1.0 12.04.2019
calcula latencias como se describe en la documentacion.
}
\examples{
# function_duracionMax(epi)
#

}
