% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_function_ValidEvents.R
\name{function_ValidEvents}
\alias{function_ValidEvents}
\title{Eventos o Episodios validos}
\usage{
function_ValidEvents(epi = NULL, drop = TRUE)
}
\arguments{
\item{epi}{Corresponde al objeto EPI}

\item{drop}{Por defecto en FALSE determina si incluye o no el "Dia 01"}
}
\value{
Devuelve una lista con los datos correctos con data management y los excluidos
}
\description{
Funcion que toma un archivo EPI con el formato del actividorm
con ligeras modificaciones y excluye algunos periodos con base en:
\enumerate{
\item Descarta el "Dia 01", aunque hay existe la opcion "drop"
\item Borra las lineas repetidas (considerar que esta funcion viene de analizar los epi antiguos)
\item Evalua que los periodos tengan al menos 3 episodios
\item Evalua que el primer episodio sea correcto
}
}
\examples{
# Calcular los antecedentes, ojo el el edit.RDS proviene del analisis hecho en la aplicacion,
# ir a esa funcion para ver como usar por fuera del programa
# setwd("D:/OneDrive/INTA/Actigrafia/testfolder")
# set <- getset(getwd())
# acvedit <- check.acvfilter("2058-001-368 JRG Baseline.AWD", set)
# filter <-  readRDS("2058-001-368 JRG Baseline.edit.RDS")
# epi <- create.epi(acvedit, filter, set)
# epi <- epi$epiviejo
#
# Y evaluar como queda
# epi <- function_ValidEvents(epi)               
# drop <- epi$drop
# epi <- epi$datos
#
}
