% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.r, R/process_settings.r
\name{getset}
\alias{getset}
\title{Leer Settings}
\usage{
getset()

getset()
}
\arguments{
\item{awdpath}{el directorio seleccionado en la app}
}
\value{
una lista con los parametros

una lista con los parametros
}
\description{
Evalua si existe o no un archivo de settings, sino crea uno y lo carga en el globalenv.

Lee el archivo de settings y carga los parametros al environment
}
