% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.awdfolder.R
\name{load.awdfolder}
\alias{load.awdfolder}
\title{Carga contenido del folder AWD}
\usage{
load.awdfolder(filelist = NULL)
}
\arguments{
\item{filelist}{vector de archivos que se obtiene con dir()}
}
\value{
data.frame con el nombre y status del archivo
}
\description{
Carga el listado de archivos AWD de un folder
}
\examples{
# awdfolder <- "D:/OneDrive/INTA/Patricio Peirano/2019.12 Kansas/kansas"
# load.awdfolder(awdfolder)#'
}
