% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.savedir.R
\name{load.savedir}
\alias{load.savedir}
\title{Cargar directorio}
\usage{
load.savedir(mainfolder)
}
\arguments{
\item{mainfolder}{Ruta}
}
\value{
El string con la ruta
}
\description{
Carga el ultimo directorio en el que se trabajo
}
