% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olitoSleep.R
\name{olitosleep}
\alias{olitosleep}
\title{Lanza la aplicacion olitosleep}
\usage{
olitosleep(inter = TRUE)
}
\arguments{
\item{inter}{Si no tiene internet pasar a FALSO para utilizar, de lo contrario se realizaran chequeos.}
}
\description{
Esta funcion carga la aplicacion de shiny.
}
\examples{
# olitosleep()
}
