% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omerge.r
\name{omerge}
\alias{omerge}
\title{omerge}
\usage{
omerge(xdf = NULL, ydf = NULL, byvar = NULL, keep = FALSE, output = TRUE)
}
\arguments{
\item{xdf}{DataFrame, con "idvar" actua de "master"}

\item{ydf}{DataFrame, con "idvar" actua de "using"}

\item{byvar}{String, Variable idvar}

\item{keep}{Logical, si guarda o no el resultado, si no solo da el reporte}
}
\value{
Lista con el merge separado en solo using, master, merge y un reporte
}
\description{
Esta funci?n es un wrapper del merge.table que hace problema cuando la base "using" solo trae una variable. Entonces esta funci?n hace todo el merge completo porque es un cacho ir contando variables mejor que sea todo automatico.
}
\examples{
# Need some data
data <- mtcars
data$cars <- row.names(mtcars)

# Prepara archivos para cruzar
# xdf <- data[-c(1:8),   c("cars", "mpg", "cyl", "disp", "hp", "drat", "wt")]
# ydf <- data[-c(26:32), c("cars", "qsec", "vs", "am", "gear", "carb")]
# byvar <- "cars"
# keep <- FALSE

# omerge(xdf, ydf, "cars")
# omerge(xdf, ydf, "cars", output = FALSE)
# omerge(xdf, ydf, "cars", keep = TRUE)
# omerge(xdf, ydf, "cars", keep = TRUE, output = FALSE)

}
