% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2datehr.R
\name{txt2datehr}
\alias{txt2datehr}
\title{Texto y Date a posix}
\usage{
txt2datehr(date, txt)
}
\arguments{
\item{date}{Fecha en formato texto}

\item{txt}{Hora en algun formato}
}
\value{
dmy_hm
}
\description{
Valida un texto y revisa si es hora en formato "hh:mm" si lo es toma el date, arrejunta y transforma en fecha
}
\examples{
# date <- "09/02/2000"
# txt <- "hh:mm"
# txt <- "23:22"
# txt <- "02:23"
# txt2datehr(date, txt)
}
